+++ neural_compressor/adaptor/torch_utils/util.py
@@ -1138,6 +1138,30 @@ def get_block_prefix(model, module_types=[torch.nn.ModuleList]):
     return block_prefix, block_num
 
 
+def calibration(model, dataloader=None, n_samples=128, calib_func=None):
+    """ Calibration with dataloader or calib_func
+
+    Args:
+        model (torch.nn.Module): input model
+        dataloader: dataloader. Defaults to None.
+        n_samples (int, optional): n_samples. Defaults to 128.
+        calib_func: calib_func. Defaults to None.
+    """
+    # calibration with dataloader or calib_func
+    if calib_func is not None:
+        calib_func(model)
+    else:
+        import math
+        from .smooth_quant import model_forward
+        batch_size = dataloader.batch_size
+        iters = int(math.ceil(n_samples / batch_size))
+        if n_samples % batch_size != 0:
+            logger.info("calibration samples increase from {} to {} due to batch_size is {}".format(
+                n_samples, iters*batch_size, batch_size,
+            ))
+        model_forward(model, dataloader, iters, next(model.parameters()).device)
+
+
 def get_hidden_states(model, dataloader=None, n_samples=128, calib_func=None):
     """get the input args and kwargs of first block.
